[refactor] 여러 곳에서 쓰이는 서버시간관련 상태를 shared feature로 분리
