+++ src/mainPage/shared/realtimeEvent/constants.js
